+++ packages/neon_framework/lib/src/utils/global_options.dart
@@ -6,6 +6,7 @@ import 'package:neon_framework/src/models/label_builder.dart';
 import 'package:neon_framework/src/settings/models/option.dart';
 import 'package:neon_framework/src/settings/models/options_collection.dart';
 import 'package:neon_framework/src/settings/models/storage.dart';
+import 'package:neon_framework/storage.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:permission_handler/permission_handler.dart';
 import 'package:universal_io/io.dart';
@@ -20,7 +21,7 @@ class GlobalOptions extends OptionsCollection {
   /// Creates a new global options collection.
   GlobalOptions(
     this._packageInfo,
-  ) : super(const AppStorage(StorageKeys.global)) {
+  ) : super(NeonStorage().settingsStorage(StorageKeys.global)) {
     pushNotificationsEnabled.addListener(_pushNotificationsEnabledListener);
     rememberLastUsedAccount.addListener(_rememberLastUsedAccountListener);
   }
