+++ packages/neon_framework/lib/src/utils/push_utils.dart
@@ -40,7 +40,7 @@ class PushUtils {
   static Future<void> Function(PushNotification notification)? onLocalNotificationClicked;
 
   static RSAKeypair loadRSAKeypair() {
-    const storage = AppStorage(StorageKeys.notifications);
+    final storage = NeonStorage().settingsStorage(StorageKeys.notifications);
     const keyDevicePrivateKey = 'device-private-key';
 
     final RSAKeypair keypair;
