+++ packages/neon_framework/lib/storage.dart
@@ -3,7 +3,7 @@
 /// The `NeonStorage` manages all storage backends.
 library;
 
-export 'package:neon_framework/src/storage/app_storage.dart';
 export 'package:neon_framework/src/storage/request_cache.dart';
+export 'package:neon_framework/src/storage/settings_store.dart' show SettingsStorage;
 export 'package:neon_framework/src/storage/single_value_store.dart';
 export 'package:neon_framework/src/storage/storage_manager.dart';
